+++ src/pages/Error-Page.tsx
@@ -1,4 +1,4 @@
-import { useRouteError } from "react-router-dom";
+import { Link, useRouteError } from "react-router-dom";
 
 export default function ErrorPage() {
   const error = useRouteError();
@@ -8,7 +8,9 @@ export default function ErrorPage() {
     <main className="container error-page">
         <p className="error-page-font">Something went wrong...</p>
         <p className="error-page-font">Go back to login page</p>
+        <Link to='/'>
         <button className="error-page-button">Click!</button>
+        </Link>
     </main>
   );
 }