+++ src/pages/Random-Recipe.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-function RandomRecipe() {
-  return (
-    <p>Hello from random recipe</p>
-  );
-}
-
-export default RandomRecipe;