+++ src/pages/Sample-Page.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-function Sample() {
-  return (
-    <p>Hello from sample</p>
-  );
-}
-
-export default Sample;