+++ changelog.md
@@ -7,3 +7,10 @@ v1.0.1:
     - Added Login page
     - Added basic pages in pages folder
     - Added axios
+v1.0.2:
+    - Created Layout component 
+    - Created Navbar, Footer component
+    - Created ApiService
+    - Created http-common with axios
+    - Created Recipe component
+    - Created RandomRecipe page